hidreport is a hid report parsing crate
This crate will eventually provide a HID report descriptor parser,
hopefully with near-zero external dependencies.

Currently working is the itemizer which is good enough to produce the
initial output of the hid-tools hid-recorder/hid-describe utilities.

Actual report parsing is incomplete but things are moving too much now
so it's better to have this all in git.

The data files are from hidtools, they are parsed with a custom script
and converted via jinja into the src/hut.rs file which provides enums
for all the values (after replacing a whole bunch of disallowed
characters). The jinja handling is mostly from libei's ei-scanner.