lib: switch from ParsedReport/ReportValue... to Field::extract
This is a much simpler implementation and turns out it's just about the
same effort for the caller. And for use-cases beyond hid-recorder
it's likely what's more commonly used anyway.

So instead of parsing a report into an InputReport and then matching
fields with values a caller can just look up the report and then call
extract_u32 for each field they actually care about.