If report size/count are unset, assume zero
This will (probably) match the behavior of C parsers where the structs
are memset on stack push/pop or main items, and thus our default value
is always zero.

Unfortunately HID report descriptors with a zero count and/or size exist
in the wild (a lot of Wacom ones do) so we have to be lenient here.