hid: default LogicalMaximum and PhysicalMaximum to unsigned
Our datatype stores it as i32 but let's interpret this as unsigned data
by default. This works around issues like 1-byte 0xff maximums for 255
which needs to be unsigne. The base problem here is that we cannot know
whether a maximum is signed or unsigned unless we look at the minimum.
But here we're just itemizing, we cannot know which one of the two is.

This is good enough though, very few devices will have negative
maximums.