types: fix UnitExponent data parsing
The HID Spec Section 6.2.2.7 lists some examples for the unit exponent
and implies that value must be a 4-bit nibble. That's not true for
real devices which will set this field just like any other value.

Let's try to parse both - any value 0xf or less is treated as a
signed 4-bit value, any other value as normal value.